refactor(auth): move auth-related hooks into seperate folder and adjust imports
